split markdownify into eight files
