+++ cs251tk/student/markdownify/truncate.py
@@ -1,0 +1,3 @@
+def truncate(string, length, encoding='utf-8'):
+    encoded = string.encode(encoding)[:length]
+    return encoded.decode(encoding, 'ignore')
